requirements.txt
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
argh==0.26.2
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
attrs==17.2.0
Automat==0.6.0
aws-shell==0.1.1
awscli==1.11.82
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
blaze==0.11.3
bleach==2.0.0
bokeh==0.12.10
boto==2.48.0
boto3==1.4.4
botocore==1.5.45
Bottleneck==1.2.1
breadability==0.1.20
bz2file==0.98
certifi==2017.11.5
cffi==1.10.0
chardet==3.0.4
chest==0.2.3
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
colorlover==0.2.1
configobj==5.0.6
constantly==15.1.0
contextlib2==0.5.5
coverage==4.4.2
cryptography==2.0.3
cssselect==1.0.1
cycler==0.10.0
cymem==1.31.2
Cython==0.26.1
cytoolz==0.8.2
dash==0.19.0
dash-core-components==0.15.0rc1
dash-html-components==0.8.0
dash-renderer==0.11.1
dash-table-experiments==0.5.0
dask==0.16.0
datashape==0.5.4
decorator==4.1.2
dill==0.2.7.1
distributed==1.20.0
docopt==0.6.2
docutils==0.14
dropbox==8.5.0
en-core-web-sm==1.2.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
ftfy==4.4.3
funcy==1.8
future==0.16.0
fuzzywuzzy==0.15.1
gensim==3.0.1
gevent==1.2.2
gmpy2==2.0.8
google-api-python-client==1.6.4
greenlet==0.4.12
h5py==2.7.1
heapdict==1.0.0
html5lib==0.999999999
httplib2==0.10.3
hyperlink==17.1.1
hypothesis==3.37.0
idna==2.6
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
inflect==0.2.5
intrinio==0.1.17
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
jellyfish==0.3.2
Jinja2==2.9.6
jmespath==0.9.2
joblib==0.11
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
language-check==1.1
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==4.1.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.8.1
mpmath==0.19
msgpack-python==0.4.8
msgraph==0.1.1
multipledispatch==0.4.9
murmurhash==0.26.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==1.11
nltk==3.2.5
nose==1.3.7
notebook==5.2.1
numba==0.35.0+6.gaa35fb1
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.7.0
oauth2client==4.1.2
odo==0.5.1
olefile==0.44
onedrivesdk==1.1.8
openpyxl==2.4.8
pandas==0.21.0
pandas-datareader==0.5.0
pandocfilters==1.4.2
parsel==1.2.0
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
pathtools==0.1.2
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pg8000==1.10.6+24.gcd4b726
pickleshare==0.7.4
Pillow==4.2.1
pipreqs==0.4.9
plac==0.9.6
plotly==2.2.1
ply==3.10
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.3.0
psutil==5.4.0
psycopg2==2.7.3.1
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0
PyDispatcher==2.0.5
PyDrive==1.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyLDAvis==2.1.1
pylint==1.7.4
Pympler==0.5
PyMySQL==0.7.11
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.5
python-dateutil==2.6.1
python-google-places==1.4.1
python-Levenshtein==0.12.0
python-onedrive==14.9.3
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
queuelib==1.4.2
redis==2.10.6
regex==2017.11.9
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
rope==0.10.7
rpy2==2.9.0
rsa==3.4.2
ruamel-yaml==0.11.14
s3transfer==0.1.10
scikit-image==0.13.1
scikit-learn==0.19.1
scikit-plot==0.3.1
scipy==1.0.0
Scrapy==1.4.0
seaborn==0.8.1
selenium==3.7.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
sortedcontainers==1.5.7
spacy==1.9.0
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
SQLAlchemy==1.1.13
squarify==0.3.0
statsmodels==0.8.0
sumy==0.7.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.1.0
termcolor==1.1.0
terminado==0.6
testpath==0.3.1
thinc==6.5.2
toolz==0.8.2
toree==0.1.0
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
Twisted==17.9.0
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.8
wrapt==1.10.11
xgboost==0.6a2
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.2.0
yarg==0.1.9
zict==0.1.3
zope.interface==4.4.3
